Adds solution for day 3 puzzles
Implements solutions for 1st part of the day 3 puzzles, with a stub for 2nd part.

Adds a helper function to validate input from the provided
aocInput.txt file.

Part 1 finds the largest digit and the second largest
digit in each input string and sums the combined
values together for all inputs.

Part 2 is currently a placeholder.